break settings into sub settings
