+++ app/src/main/res/values/styles.xml
@@ -41,10 +41,10 @@
     </style>
 
     <style name="PreferenceCategory">
-        <item name="android:layout_marginTop">2dp</item>
-        <item name="android:layout_marginBottom">2dp</item>
+        <item name="android:layout_marginTop">8dp</item>
+        <item name="android:layout_marginBottom">8dp</item>
         <item name="paddingStart">10dp</item>
-        <item name="android:textSize">20sp</item>
+        <item name="android:textSize">24sp</item>
         <item name="android:textColor">#000000</item>
     </style>
